Prioritise indentation from CLI (#4)
* Prioritise indentation from CLI

When the indentation is set from the CLI arguments, respect it.
Otherwise, and by default, use the indentation from the editorconfig
settings, if any. This default behaviour is controlled through a
negative indentation. Whenever no indentation can be found, default to
the constant in the implementation, i.e. 2 (matches the one in lxml).

* Refactor .editorconfig handling

Move handling of .editorconfig, including default values to a separate
function. Use the new function in the main 'beautify' funtion.

* Get indentation from .editorconfig on strict neg.

Indentation settings are picked from the .editorconfig file only when
the CLI argument is strictly negative. This is to enable no indentation
of the XML file, i.e. when indentation is set to `0` exactly.