Better control of line endings (#10)
* Add line endings "intelligence"

Add a command-line option and associated environment variable to control
the line endings used in the formatted files. This is because the pretty
printing functionality built in LXML always uses LF line endings (unix).
The implementation can use 'unix', 'windows' and 'mac' (classic) line
endings. There is also a mode called 'auto', which will detect the line
endings of the original file and apply it on the result. The 'auto' mode
will not work deterministically on files that have MIXED line endings.
But this is ok, as such files should probably be sanitised anyway.

* Document new --line-endings CLI option