Add and document self-closing tags format control (#12)
Add an option (and environment variable) to control the format of
self-closing tags. When the option/var is `space` an extra space will be
added before the ending `/>` for extra readability. A mode called `auto`
will detect if a string ` />` appears in the original file and
automatically add space -- e.g. behave as `space`. `nospace` is as the
default of LXML, i.e. no extra space.

Adding a space is made the default, this matches the default of the
prettier XML plugin.

Note: the implementation of the `auto` mode is not entirely correct, but
is deemed enough for detection purposes.