Add control of the XML declaration (#18)
* Add control of the XML declaration

Add a command-line option and environment variable to control the XML
declaration. This allows to completely remove it, but also (most
usually) to ensure its quoting style. This is because the underlying XML
library will sometimes use single quotes, while the rest of the content
is using double quotes. Apart for providing an option to enforce the
style, the implementation is also able to detect the style of the
reformatted output and apply it to the declaration. The heuristic
prefers double quotes.

* Improve code quality and fix one possible bug

Improve code quality by making it more concise and fixing copy/paste
mistakes. Fix a bug on the find() method that returns a -1 value when
not found.

* Improve code further

Using not in is more idiomatic.

* Fix bullet style